requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
bson==0.5.8
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
dataclasses==0.6
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.2
django-clear-cache==0.3
django-common-helpers==0.9.2
django-cors-headers==3.2.1
django-cron==0.5.1
django-extensions==2.2.6
django-import-export==2.0
django-rq==2.2.0
django-task==1.5.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
djongo==1.3.0
et-xmlfile==1.0.1
httpie==1.0.3
idna==2.8
jdcal==1.4.1
kiwisolver==1.1.0
kombu==4.6.7
MarkupPy==1.14
matplotlib==3.1.2
numpy==1.18.1
odfpy==1.4.0
openpyxl==3.0.3
pandas==0.25.3
Pillow==7.0.0
Pygments==2.5.2
pymongo==3.10.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
redis==3.3.11
requests==2.22.0
rq==1.2.0
scipy==1.4.1
seaborn==0.9.0
six==1.13.0
sqlparse==0.2.4
tablib==0.14.0
urllib3==1.25.7
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
